requirements.txt
selenium>=4.0.0
Pillow>=8.0.0
pyvis>=0.1.9
